Renamed, updated paths, fixed deserializer
